+++ backend/reply/viewsets.py
@@ -8,7 +8,7 @@ from .models import Reply
 
 
 class ReplyViewSet(viewsets.ModelViewSet):
-    @list_route(url_path='new', methods=['post'])  # , permission_classes=[IsAuthenticated])
+    @list_route(url_path='new', methods=['post'], permission_classes=[IsAuthenticated])
     def write_reply(self, request):
         req_data = request.data
         plane_author_id = req_data['plane_author']
@@ -29,16 +29,27 @@ class ReplyViewSet(viewsets.ModelViewSet):
 
         return Response(status=status.HTTP_201_CREATED)
 
-    @list_route(url_path="(?P<reply_id>[0-9]+)")#, permission_classes=[IsAuthenticated])
+    @list_route(url_path="(?P<reply_id>[0-9]+)", permission_classes=[IsAuthenticated])
     def get_reply(self, request, reply_id):
-        reply = Reply.objects.get(id=reply_id)
+        try:
+            reply = Reply.objects.get(id=reply_id)
+        except Reply.DoesNotExist:
+            return Response(status=status.HTTP_404_NOT_FOUND)
+
+        if reply.plane_author_id != request.user.id:
+            return Response(status=status.HTTP_403_FORBIDDEN)
+
         result = model_to_dict(reply)
         result['reply_id'] = result.pop('id')
         return Response(result)
 
-    @list_route(url_path="user/(?P<user_id>[0-9]+)")#, permission_classes=[IsAuthenticated])
+    @list_route(url_path="user/(?P<user_id>[0-9]+)", permission_classes=[IsAuthenticated])
     def get_reply_by_user(self, request, user_id):
+        if int(user_id) != request.user.id:
+            return Response([])
+
         user = user_model.User.objects.get(id=user_id).user
+
         replies = user.replies
         result = []
         for reply in replies.values():
@@ -46,7 +57,7 @@ class ReplyViewSet(viewsets.ModelViewSet):
             result.append(reply)
         return Response(result)
 
-    @list_route(url_path="report", methods=['put'])#, permission_classes=[IsAuthenticated])
+    @list_route(url_path="report", methods=['put'], permission_classes=[IsAuthenticated])
     def report_reply(self, request):
         req_data = request.data
         reply_id = req_data['reply_id']
@@ -54,13 +65,17 @@ class ReplyViewSet(viewsets.ModelViewSet):
             reply = Reply.objects.get(id=reply_id)
         except Reply.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
+
+        if reply.plane_author_id != request.user.id:
+            return Response(status=status.HTTP_403_FORBIDDEN)
+
         reply.is_reported = True
         reply.save()
         reply.reply_author.decrease_likes()
         reply.reply_author.save()
         return Response(status=status.HTTP_200_OK)
 
-    @list_route(url_path="like", methods=['put'])#, permission_classes=[IsAuthenticated])
+    @list_route(url_path="like", methods=['put'], permission_classes=[IsAuthenticated])
     def like_reply(self, request):
         req_data = request.data
         reply_id = req_data['reply_id']
@@ -68,6 +83,9 @@ class ReplyViewSet(viewsets.ModelViewSet):
             reply = Reply.objects.get(id=reply_id)
         except Reply.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
+
+        if reply.plane_author_id != request.user.id:
+            return Response(status=status.HTTP_403_FORBIDDEN)
 
         if reply.liked:
             return Response(status=status.HTTP_406_NOT_ACCEPTABLE)
@@ -79,13 +97,17 @@ class ReplyViewSet(viewsets.ModelViewSet):
         reply.reply_author.save()
         return Response(status=status.HTTP_200_OK)
 
-    @list_route(url_path="delete", methods=['put'])  # , permission_classes=[IsAuthenticated])
+    @list_route(url_path="delete", methods=['put'], permission_classes=[IsAuthenticated])
     def delete_reply(self, request):
         req_data = request.data
-        plane_id = req_data['reply_id']
+        reply_id = req_data['reply_id']
         try:
-            reply = Reply.objects.get(id=plane_id)
+            reply = Reply.objects.get(id=reply_id)
         except Reply.DoesNotExist:
             return Response(status=status.HTTP_404_NOT_FOUND)
+
+        if reply.plane_author_id != request.user.id:
+            return Response(status=status.HTTP_403_FORBIDDEN)
+
         reply.delete()
         return Response(status=status.HTTP_200_OK)
